feat(attribute): refactor attribute popups
