+++ .babelrc
@@ -10,7 +10,5 @@
 		],
 		"@babel/preset-react"
 	],
-	"plugins": [
-		"@babel/plugin-proposal-class-properties"
-	]
+	"plugins": ["@babel/plugin-proposal-class-properties", "@babel/plugin-syntax-dynamic-import"]
 }
