+++ package.json
@@ -136,7 +136,8 @@
 	"devDependencies": {
 		"@babel/core": "7.1.6",
 		"@babel/plugin-proposal-class-properties": "7.1.0",
-		"@babel/plugin-syntax-dynamic-import": "7.2.0",
+		"@babel/plugin-syntax-dynamic-import": "^7.2.0",
+		"@babel/plugin-transform-runtime": "^7.6.2",
 		"@babel/preset-env": "7.1.6",
 		"@babel/preset-react": "7.0.0",
 		"@commitlint/cli": "7.2.1",
@@ -148,6 +149,7 @@
 		"@storybook/addon-links": "5.0.11",
 		"@storybook/addons": "5.0.11",
 		"@storybook/react": "5.0.11",
+		"acorn": "6.1.1",
 		"babel-core": "7.0.0-bridge.0",
 		"babel-eslint": "10.0.1",
 		"babel-jest": "24.5.0",
