+++ src/common/config.js
@@ -1,13 +1,15 @@
 /* istanbul ignore file */
 'use strict';
 const path = require('path');
-const dotenv = require('dotenv');
-const electron = require('electron');
-
+let electron;
+if (!process.env.STORYBOOK) {
+	const dotenv = require('dotenv');
+	dotenv.config();
+	electron = require('electron');
+}
 const { isDevMode, isTestMode, getSetupFilePath, getUserDataPath } = require('./utils/common');
 const pkg = require('../../package.json');
 
-dotenv.config();
 const DEBUG_REQUEST = process.env.DEBUG_REQUEST === '1';
 if (DEBUG_REQUEST) {
 	require('request').debug = true;
@@ -37,7 +39,7 @@ const DEPOSIT_PRICE_OVERRIDE = process.env.DEPOSIT_PRICE_OVERRIDE;
 
 let userDataDirectoryPath = '';
 let walletsDirectoryPath = '';
-if (electron.app) {
+if (electron && electron.app) {
 	userDataDirectoryPath = electron.app.getPath('userData');
 	walletsDirectoryPath = path.resolve(userDataDirectoryPath, 'wallets');
 }
