+++ src/renderer/attributes/delete-attribute.jsx
