+++ src/renderer/attributes/index.js
@@ -1,0 +1,3 @@
+export { EditAttributeContainer } from './edit-attribute-container';
+export { DeleteAttributeContainer } from './delete-attribute-container';
+export { CreateAttributeContainer } from './create-attribute-container';
