+++ src/renderer/attributes/transform-errors.js
